requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
numpy==1.24.3
pandas==2.0.3
tensorflow==2.14.0
scikit-learn==1.3.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
